feat: finish services and repositories for groups and deposits

+++ springboot-app/src/main/java/com/lumos/lumosspring/stock/service/DepositService.java
@@ -5,22 +5,26 @@ import com.lumos.lumosspring.stock.entities.Deposit;
 import com.lumos.lumosspring.stock.entities.Type;
 import com.lumos.lumosspring.stock.repository.CompanyRepository;
 import com.lumos.lumosspring.stock.repository.DepositRepository;
+import com.lumos.lumosspring.stock.repository.MaterialRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
 import java.util.List;
 
 @Service
 public class DepositService {
     @Autowired
     private DepositRepository depositRepository;
-
     @Autowired
     private CompanyRepository comapanyRepository;
+    @Autowired
+    private MaterialRepository materialRepository;
 
     public List<Deposit> findAll() {
-        return depositRepository.findAll();
+        return depositRepository.findAllByOrderByIdDeposit();
     }
 
     public Deposit findById(Long id) {
@@ -28,12 +32,16 @@ public class DepositService {
     }
 
     public ResponseEntity<?> save(DepositDTO depositDTO) {
+        if (depositRepository.existsByDepositName(depositDTO.depositName())) {
+            return ResponseEntity.status(HttpStatus.CONFLICT).body(Collections.singletonMap("message", "Este almoxarifado já existe."));
+        }
+
         var deposit = new Deposit();
         deposit.setDepositName(depositDTO.depositName());
         deposit.setCompany(comapanyRepository.findById(depositDTO.companyId()).orElse(null));
         depositRepository.save(deposit);
 
-        return ResponseEntity.ok(depositRepository.findAll());
+        return ResponseEntity.ok(this.findAll());
     }
 
     public ResponseEntity<?> update(Long depositId, DepositDTO depositDTO) {
@@ -50,7 +58,7 @@ public class DepositService {
         deposit.setCompany(company);
         depositRepository.save(deposit);
 
-        return ResponseEntity.ok(depositRepository.findAll());
+        return ResponseEntity.ok(this.findAll());
     }
 
     public ResponseEntity<?> delete(Long id) {
@@ -58,8 +66,13 @@ public class DepositService {
         if (type == null) {
             return ResponseEntity.notFound().build();
         }
+
+        if (materialRepository.existsDeposit(id).isPresent()) {
+            return ResponseEntity.badRequest().body(Collections.singletonMap("message", "Não é possível excluir: há materiais associados a este almoxarifado."));
+        }
+
         depositRepository.delete(type);
-        return ResponseEntity.ok(depositRepository.findAll());
+        return ResponseEntity.ok(this.findAll());
     }
 
 }
